compiler: extract context from simulate
Context should not depend on simulate. We can use context
without simulation.
